+++ src/const.js
@@ -42,4 +42,18 @@ const MAX_DAY_LIMIT = 7;
 const MIN_TAGS_COUNT = 0;
 const MAX_TAGS_COUNT = 3;
 
-export {COLORS, DAYS, MONTHS, Filters, ALL_TASKS_COUNT, ONE_TASKS_PAGE_COUNT, RANDOM_LIMIT, MIN_RANDOM_NUMBER, DIRECTION_POSITIVE, DIRECTION_NEGATIVE, MIN_DAY_LIMIT, MAX_DAY_LIMIT, MIN_TAGS_COUNT, MAX_TAGS_COUNT};
+const DESCRIPTION_ITEMS = [
+  `Изучить теорию`,
+  `Сделать домашку`,
+  `Пройти интенсив на соточку`,
+];
+
+const TAGS = [
+  `homework`,
+  `theory`,
+  `practice`,
+  `intensive`,
+  `keks`
+];
+
+export {COLORS, DAYS, MONTHS, Filters, ALL_TASKS_COUNT, ONE_TASKS_PAGE_COUNT, RANDOM_LIMIT, MIN_RANDOM_NUMBER, DIRECTION_POSITIVE, DIRECTION_NEGATIVE, MIN_DAY_LIMIT, MAX_DAY_LIMIT, MIN_TAGS_COUNT, MAX_TAGS_COUNT, DESCRIPTION_ITEMS, TAGS};
