+++ src/mock/task.js
@@ -1,10 +1,5 @@
-import {COLORS, RANDOM_LIMIT, MIN_RANDOM_NUMBER, DIRECTION_POSITIVE, DIRECTION_NEGATIVE, MIN_DAY_LIMIT, MAX_DAY_LIMIT, MIN_TAGS_COUNT, MAX_TAGS_COUNT} from '../const.js';
-
-const DescriptionItems = [
-  `Изучить теорию`,
-  `Сделать домашку`,
-  `Пройти интенсив на соточку`,
-];
+import {COLORS, MIN_TAGS_COUNT, MAX_TAGS_COUNT, DESCRIPTION_ITEMS, TAGS} from '../const.js';
+import {getRandomBoolean, getRandomDate, getRandomItem} from '../utils.js';
 
 const DefaultRepeatingDays = {
   'mo': false,
@@ -16,43 +11,8 @@ const DefaultRepeatingDays = {
   'su': false,
 };
 
-const Tags = [
-  `homework`,
-  `theory`,
-  `practice`,
-  `intensive`,
-  `keks`
-];
-
-const getRandomNumber = (min, max) => {
-  return Math.floor((max - min) * Math.random());
-};
-
-const getRandomBoolean = () => {
-  return Math.random() > RANDOM_LIMIT;
-};
-
-const getRandomDate = () => {
-  const currentDate = new Date();
-
-  const directionDay = getRandomBoolean() ? DIRECTION_POSITIVE : DIRECTION_NEGATIVE;
-  const day = directionDay * getRandomNumber(MIN_DAY_LIMIT, MAX_DAY_LIMIT);
-
-  currentDate.setDate(currentDate.getDate() + day);
-
-  return currentDate;
-};
-
-const getRandomItem = (items) => {
-  const index = getRandomNumber(MIN_RANDOM_NUMBER, items.length);
-  return items[index];
-};
-
-
 const createTags = (tags, min, max) => {
-  return tags
-    .filter(() => getRandomBoolean())
-    .slice(min, max);
+  return tags.filter(getRandomBoolean).slice(min, max);
 };
 
 const createRepeatingDays = () => {
@@ -63,10 +23,10 @@ const createTask = () => {
   const dueDate = getRandomBoolean() ? null : getRandomDate();
 
   return {
-    description: getRandomItem(DescriptionItems),
+    description: getRandomItem(DESCRIPTION_ITEMS),
     dueDate,
     repeatingDays: dueDate !== null ? DefaultRepeatingDays : createRepeatingDays(),
-    tags: new Set(createTags(Tags, MIN_TAGS_COUNT, MAX_TAGS_COUNT)),
+    tags: new Set(createTags(TAGS, MIN_TAGS_COUNT, MAX_TAGS_COUNT)),
     color: getRandomItem(COLORS),
     isFavorite: getRandomBoolean(),
     isArchive: getRandomBoolean(),
